strategic_engine/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pinecone-client==2.2.4
openai==1.3.7
PyPDF2==3.0.1
python-docx==0.8.11
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1